examples/subobc: apply rename migration(Applications -> applications)
