+++ examples/subobc/src/src_user/Settings/Modes/TaskLists/Elements/tl_elem_gs_related_process.c
@@ -2,7 +2,7 @@
 #include "tl_elem_gs_related_process.h"
 
 #include "../../../../TlmCmd/block_command_definitions.h"
-#include "../../../../Applications/app_registry.h"
+#include "../../../../applications/app_registry.h"
 
 #include <src_core/TlmCmd/block_command_loader.h>
