+++ examples/subobc/src/src_user/TlmCmd/command_source.h
@@ -17,6 +17,6 @@
 #include <src_core/TlmCmd/command_dispatcher_manager.h>
 #include <src_core/TlmCmd/telemetry_generator.h>
 #include <src_core/TlmCmd/telemetry_frame.h>
-#include "../Applications/app_headers.h"
+#include "../applications/app_headers.h"
 
 #endif
