+++ examples/subobc/src/src_user/TlmCmd/telemetry_source.h
@@ -17,7 +17,7 @@
 #include <src_core/TlmCmd/command_analyze.h>
 #include <src_core/TlmCmd/command_dispatcher_manager.h>
 #include <src_core/TlmCmd/block_command_executor.h>
-#include "../Applications/app_headers.h"
+#include "../applications/app_headers.h"
 #include "../Settings/build_settings.h"
 
 #endif
